Project _5
The working of this project is based on Reactjs, Unsplash Api, and Axios Library. This project asks users what they want to search after getting user input it takes it and makes an HTTP request to Unsplash API after getting the response From Unsplash API, it shows the images which are associated with the user search. This project tells us about API how to use them ., how to do an HTTP request, and how to read user input using the STATE system.



You have to install node-module as well as axios library
npm install axios